+++ resteasy-spring-boot-starter/src/main/java/com/paypal/springboot/resteasy/ResteasyEmbeddedServletInitializer.java
@@ -31,8 +31,15 @@ import java.util.Set;
  */
 public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProcessor {
 
-    private static final String JAXRS_APP_CLASSES_PROPERTY = "resteasy.jaxrs.app";
     private static final String JAXRS_APP_CLASSES_DEFINITION_PROPERTY = "resteasy.jaxrs.app.registration";
+    private static final String JAXRS_APP_CLASSES_PROPERTY = "resteasy.jaxrs.app.classes";
+
+    // This is how JAXRS_APP_CLASSES_PROPERTY was named originally. It conflicted with "resteasy.jaxrs.app.registration"
+    // in case of YAML files, since registration was a child of app from an YAML perspective, which is not allowed.
+    // Because of that its name was changed (the ".classes" suffix was added).
+    // This legacy property has not been removed though, to keep backward compatibility, but it is marked as deprecated. It will be
+    // available only for .properties files, but not for YAML files. It should be finally removed in a future major release.
+    private static final String JAXRS_APP_CLASSES_PROPERTY_LEGACY = "resteasy.jaxrs.app";
 
     private Set<Class<? extends Application>> applications = new HashSet<Class<? extends Application>>();
     private Set<Class<?>> allResources = new HashSet<Class<?>>();
@@ -49,13 +56,13 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
      * This is done by one of these three options in this order:
      *
      * 1- By having them defined as Spring beans
-     * 2- By setting property {@code resteasy.jaxrs.app} via Spring Boot application properties file.
+     * 2- By setting property {@code resteasy.jaxrs.app.classes} via Spring Boot application properties file.
      *    This property should contain a comma separated list of JAX-RS sub-classes
      * 3- Via classpath scanning (looking for javax.ws.rs.core.Application sub-classes)
      *
      * First try to find JAX-RS Application sub-classes defined as Spring beans. If that is existent,
      * the search stops, and those are the only JAX-RS applications to be registered.
-     * If no JAX-RS application Spring beans are found, then see if Sring Boot property {@code resteasy.jaxrs.app}
+     * If no JAX-RS application Spring beans are found, then see if Spring Boot property {@code resteasy.jaxrs.app.classes}
      * has been set. If it has, the search stops, and those are the only JAX-RS applications to be registered.
      * If not, then scan the classpath searching for JAX-RS applications.
      *
@@ -139,14 +146,18 @@ public class ResteasyEmbeddedServletInitializer implements BeanFactoryPostProces
     }
 
     /**
-     * Find JAX-RS application classes via property {@code resteasy.jaxrs.app}
+     * Find JAX-RS application classes via property {@code resteasy.jaxrs.app.classes}
      */
     private void findJaxrsApplicationProperty(ConfigurableListableBeanFactory beanFactory) {
         ConfigurableEnvironment configurableEnvironment = beanFactory.getBean(ConfigurableEnvironment.class);
         String jaxrsAppsProperty = configurableEnvironment.getProperty(JAXRS_APP_CLASSES_PROPERTY);
         if(jaxrsAppsProperty == null) {
-            logger.info("No JAX-RS Application set via property {}", JAXRS_APP_CLASSES_PROPERTY);
-            return;
+            jaxrsAppsProperty = configurableEnvironment.getProperty(JAXRS_APP_CLASSES_PROPERTY_LEGACY);
+            if(jaxrsAppsProperty == null) {
+                logger.info("No JAX-RS Application set via property {}", JAXRS_APP_CLASSES_PROPERTY);
+                return;
+            }
+            logger.warn("Property {} has been set. Notice that this property has been deprecated and will be removed soon. Please replace it by property {}", JAXRS_APP_CLASSES_PROPERTY_LEGACY, JAXRS_APP_CLASSES_PROPERTY);
         } else {
             logger.info("Property {} has been set", JAXRS_APP_CLASSES_PROPERTY);
         }