+++ resteasy-spring-boot-starter/src/test/java/com/paypal/springboot/resteasy/JaxrsAppRegistrationTest.java
@@ -24,7 +24,8 @@ import java.util.*;
 public class JaxrsAppRegistrationTest {
 
     private static final String DEFINITION_PROPERTY = "resteasy.jaxrs.app.registration";
-    private static final String APP_CLASSES_PROPERTY = "resteasy.jaxrs.app";
+    private static final String APP_CLASSES_PROPERTY = "resteasy.jaxrs.app.classes";
+    private static final String APP_CLASSES_PROPERTY_LEGACY = "resteasy.jaxrs.app";
 
     private static Set<Class> allPossibleAppClasses;
 
@@ -86,6 +87,19 @@ public class JaxrsAppRegistrationTest {
         ConfigurableEnvironment configurableEnvironmentMock = mock(ConfigurableEnvironment.class);
         when(configurableEnvironmentMock.getProperty(DEFINITION_PROPERTY)).thenReturn("property");
         when(configurableEnvironmentMock.getProperty(APP_CLASSES_PROPERTY)).thenReturn("com.paypal.springboot.resteasy.sample.TestApplication3, com.paypal.springboot.resteasy.sample.TestApplication4,com.paypal.springboot.resteasy.sample.TestApplication2");
+
+        Set<Class> expectedRegisteredAppClasses = new HashSet<Class>();
+        expectedRegisteredAppClasses.add(TestApplication2.class);
+        expectedRegisteredAppClasses.add(TestApplication4.class);
+
+        test(configurableEnvironmentMock, expectedRegisteredAppClasses);
+    }
+
+    @Test
+    public void legacyPropertyTest() {
+        ConfigurableEnvironment configurableEnvironmentMock = mock(ConfigurableEnvironment.class);
+        when(configurableEnvironmentMock.getProperty(DEFINITION_PROPERTY)).thenReturn("property");
+        when(configurableEnvironmentMock.getProperty(APP_CLASSES_PROPERTY_LEGACY)).thenReturn("com.paypal.springboot.resteasy.sample.TestApplication3, com.paypal.springboot.resteasy.sample.TestApplication4,com.paypal.springboot.resteasy.sample.TestApplication2");
 
         Set<Class> expectedRegisteredAppClasses = new HashSet<Class>();
         expectedRegisteredAppClasses.add(TestApplication2.class);